expense-tracker-backend/requirement.txt
fastapi[all]
uvicorn
passlib[bcrypt]
python-jose
python-multipart
pandas
numpy
scikit-learn
anyio
email-validator
h11
pydantic
annotated-types
pydantic-core
pymongo
sniffio
motor
python-dotenv
PyJWT

